Don't break older versions of Dockstore CLI (#3)
ga4gh/dockstore#1629

One half of fix to not break older versions of Dockstore CLI when
this plugin is present in the ~/.dockstore/plugins directory.

In version 1.4 and earlier of Dockstore CLI
`VersionAwarePluginManager.getSystemVersion()` returns a Version
that represents `0.0.0`, the default; by specifying that this
plugin version requires >0.0.0. An upcoming PR to Dockstore Client
will change that version to 0.0.1.

When Dockstore CLI 1.4 runs, it will not attempt to load this plugin,
because of the requires requirement.

Fortunately, we are using an older version of the PF4J, as the
more recent versions interpret the 0.0.0 system version to mean
load all plugins without checking the version.